Bi-C2R/requirement.txt
Cython
easydict
h5py
huggingface-hub
imageio
matplotlib
metric-learn
multidict
numpy
opencv-python
pandas
Pillow
pytorch-lightning
PyYAML
scikit-image
scikit-learn
scipy
tensorboard
tensorflow
timm
safetensors==0.4.1
tokenizers==0.13.0
tqdm
yacs
prettytable

